[MARK-018] View 관련 코드 View Package 하위로 이동
