+++ DevMark/app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
         </activity>
 
         <activity
-            android:name=".presentation.SignIn.SignInActivity"
+            android:name=".presentation.view.SignIn.SignInActivity"
             android:exported="true" />
 
         <activity
@@ -36,7 +36,7 @@
             android:exported="true" />
 
         <activity
-            android:name=".presentation.WorkspaceSelect.SelectWorkspaceActivity"
+            android:name=".presentation.view.WorkspaceSelect.SelectWorkspaceActivity"
             android:exported="true" />
 
         <activity
