+++ DevMark/app/src/main/java/com/devmark/devmark/MainActivity.kt
@@ -8,10 +8,10 @@ import android.view.View
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import com.devmark.devmark.databinding.ActivityMainBinding
-import com.devmark.devmark.presentation.Recommend.RecommendFragment
-import com.devmark.devmark.presentation.Remind.RemindFragment
-import com.devmark.devmark.presentation.Workspace.WorkspaceFragment
-import com.devmark.devmark.presentation.WorkspaceSelect.SelectWorkspaceActivity
+import com.devmark.devmark.presentation.view.Recommend.RecommendFragment
+import com.devmark.devmark.presentation.view.Remind.RemindFragment
+import com.devmark.devmark.presentation.view.Workspace.WorkspaceFragment
+import com.devmark.devmark.presentation.view.WorkspaceSelect.SelectWorkspaceActivity
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
