--- DevMark/app/src/main/java/com/devmark/devmark/presentation/WorkspaceSelect/WorkspaceAddDialog.kt
@@ -1,5 +1,0 @@
-package com.devmark.devmark.presentation.WorkspaceSelect
-
-class WorkspaceAddDialog {
-
-}
