+++ DevMark/app/src/main/java/com/devmark/devmark/presentation/view/Recommend/TopCategoryLvAdapter.kt
@@ -1,4 +1,4 @@
-package com.devmark.devmark.presentation.Recommend
+package com.devmark.devmark.presentation.view.Recommend
 
 import android.annotation.SuppressLint
 import android.content.Context
