+++ DevMark/app/src/main/java/com/devmark/devmark/presentation/view/Workspace/OnItemClickListener.kt
@@ -1,4 +1,4 @@
-package com.devmark.devmark.presentation.Workspace
+package com.devmark.devmark.presentation.view.Workspace
 
 interface OnItemClickListener {
     fun onClick(name: String)
