+++ DevMark/app/src/main/java/com/devmark/devmark/presentation/view/Workspace/WorkSpaceRvAdapter.kt
@@ -1,4 +1,4 @@
-package com.devmark.devmark.presentation.Workspace
+package com.devmark.devmark.presentation.view.Workspace
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
