+++ DevMark/app/src/main/java/com/devmark/devmark/presentation/view/Workspace/WorkspaceFragment.kt
@@ -1,4 +1,4 @@
-package com.devmark.devmark.presentation.Workspace
+package com.devmark.devmark.presentation.view.Workspace
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,8 +10,8 @@ import com.devmark.devmark.MainActivity
 import com.devmark.devmark.R
 import com.devmark.devmark.data.BookMark
 import com.devmark.devmark.databinding.FragmentWorkspaceBinding
-import com.devmark.devmark.presentation.Setting.SettingFragment
-import com.devmark.devmark.presentation.WorkspaceSetting.WorkspaceSettingFragment
+import com.devmark.devmark.presentation.view.Setting.SettingFragment
+import com.devmark.devmark.presentation.view.WorkspaceSetting.WorkspaceSettingFragment
 
 class WorkspaceFragment : Fragment() {
     private lateinit var binding: FragmentWorkspaceBinding
