+++ DevMark/app/src/main/java/com/devmark/devmark/presentation/view/WorkspaceSelect/WorkspaceAddDialog.kt
@@ -1,0 +1,5 @@
+package com.devmark.devmark.presentation.view.WorkspaceSelect
+
+class WorkspaceAddDialog {
+
+}
